特性：自动备份插件增加下载判断

+++ plugin/auto-backup.sh
@@ -14,6 +14,11 @@ Install() {
     rm -rf /www/panel/plugins/AutoBackup
     mkdir /www/panel/plugins/AutoBackup
     wget -O /www/panel/plugins/AutoBackup/auto-backup.zip "https://api.panel.haozi.xyz/api/plugin/url?slug=auto-backup"
+    if [ "$?" != "0" ]; then
+        echo -e $HR
+        echo "错误：下载失败，请检查网络是否正常。"
+        exit 1
+    fi
     cd /www/panel/plugins/AutoBackup
     unzip -o auto-backup.zip && rm -rf auto-backup.zip
     # 写入插件安装状态
